Making YTArray and YTQuantity default to dtype=float64.
This fixes issues with in-place conversions where the dtype is initially an int.

For example:

arr = YTArray([1,2,3], 'cm')

will have a dtype of int64.  This is fine until we convert to a different unit
that has a non-integer conversion factor.  If this happens, numpy will cast
to int64 since in-place ufuncs cannot alter the dtype of an array.

Defaulting to float64 avoids this issue while still allowing users to select a
different dtype if they really want to.

--HG--
branch : yt-3.0