give unit registries a unit system
this required a fair amount of refactoring to avoid a circular dependency
generated by the fact that the unit systems stored Unit objects. That's not
actually necessary, the unit systems only need to store the units that make up
the system and can generate Unit objects to return to the user when __getitem__
is called.

Now that unit registries are associated with a unit system, the registry can be
used to determine the appropriate unit system to use for a calculation. This
allows the unit system to be controlled by an application by making use of a
custom unit registry. Calls to in_base or get_base_equivalent will return data
in the appropriately configured unit system associated with that registry.