Allow == and != for data with different units. Fixes #105.
Up until now we were unconditionally raising a UnitOperationError for all
comparisons of data with different units, including `equal` and `not_equal`. Due
to details of how NumPy handles this internally (I guess it happens in C code
because I don't see any of this when stepping through in pdb) numpy swallows
that error and raises a DeprecationWarning (!?) causing confusing behavior when
comparing quantities with units that are not dimensionally equivalent.

The fix is to add special handling for equal and not_equal when comparing data
with units that have different dimensions. I've also added tests for this case.

This may have negative performance implications for some workflows however I
think we have to do this in the interest of correctness. Also comparing data
with different units like this is kinda a weird thing to do and hopefully won't
impact common workflows.