Update tests to reflect changes in CODATA 2022 constant values
Changing the number of decimal places in kB (to the CODATA value) meant that assert_equals wasn't working. So I have changed it to assert_allclose, which is consistent with every other comparison in the same block. This meant writing and using the `assert_lists_allclose` function from 87fe8f17fc6b02dfb3c9adf1113b882feb50795f.

This commit fails because of electromagnetic units. This is something I'm working on.